Improvements?

+++ project/include/time_ctrl.h
@@ -3,6 +3,6 @@
 
 int getStartTime();
 
-double getMillisecondsElapsed();
+int getMillisecondsElapsed();
 
 #endif // PROJECT_INCLUDE_TIME_CTRL_H_
